+++ src/shared/components/results/ResultsView.tsx
@@ -18,9 +18,7 @@ import DiseasesCard from '../../../supporting-data/diseases/components/results/D
 import DatabaseCard from '../../../supporting-data/database/components/results/DatabaseCard';
 import LocationsCard from '../../../supporting-data/locations/components/results/LocationsCard';
 
-import uniProtKbConverter, {
-  UniProtkbAPIModel,
-} from '../../../uniprotkb/adapters/uniProtkbConverter';
+import { UniProtkbAPIModel } from '../../../uniprotkb/adapters/uniProtkbConverter';
 import { UniRefLiteAPIModel } from '../../../uniref/adapters/uniRefConverter';
 import { UniParcAPIModel } from '../../../uniparc/adapters/uniParcConverter';
 import { ProteomesAPIModel } from '../../../proteomes/adapters/proteomesConverter';
@@ -31,24 +29,12 @@ import { DiseasesAPIModel } from '../../../supporting-data/diseases/adapters/dis
 import { DatabaseAPIModel } from '../../../supporting-data/database/adapters/databaseConverter';
 import { LocationsAPIModel } from '../../../supporting-data/locations/adapters/locationsConverter';
 
-// columns for table views
-import UniProtKBColumnConfiguration from '../../../uniprotkb/config/UniProtKBColumnConfiguration';
-import UniRefColumnConfiguration from '../../../uniref/config/UniRefColumnConfiguration';
-import UniParcColumnConfiguration from '../../../uniparc/config/UniParcColumnConfiguration';
-import ProteomesColumnConfiguration from '../../../proteomes/config/ProteomesColumnConfiguration';
-import TaxonomyColumnConfiguration from '../../../supporting-data/taxonomy/config/TaxonomyColumnConfiguration';
-import KeywordsColumnConfiguration from '../../../supporting-data/keywords/config/KeywordsColumnConfiguration';
-import CitationsColumnConfiguration from '../../../supporting-data/citations/config/CitationsColumnConfiguration';
-import DiseasesColumnConfiguration from '../../../supporting-data/diseases/config/DiseasesColumnConfiguration';
-import DatabaseColumnConfiguration from '../../../supporting-data/database/config/DatabaseColumnConfiguration';
-import LocationsColumnConfiguration from '../../../supporting-data/locations/config/LocationsColumnConfiguration';
-
 import useDataApi from '../../hooks/useDataApi';
 import useNS from '../../hooks/useNS';
 import useNSQuery from '../../hooks/useNSQuery';
 import usePrefetch from '../../hooks/usePrefetch';
 import useUserPreferences from '../../hooks/useUserPreferences';
-import useColumnSort from '../../hooks/useColumnSort';
+import useColumns, { ColumnDescriptor } from '../../hooks/useColumns';
 
 import fieldsForUniProtKBCards from '../../../uniprotkb/config/UniProtKBCardConfiguration';
 
@@ -58,60 +44,11 @@ import { getEntryPathFor } from '../../../app/config/urls';
 
 import { Namespace } from '../../types/namespaces';
 import { APIModel } from '../../types/apiModel';
-import { SortDirection } from '../../../uniprotkb/types/resultsTypes';
-import { SortableColumn } from '../../../uniprotkb/types/columnTypes';
-import { Column, nsToDefaultColumns } from '../../config/columns';
 import { ViewMode } from './ResultsContainer';
+import { Column } from '../../config/columns';
 
 import './styles/warning.scss';
 import './styles/results-view.scss';
-
-const convertRow = (row: APIModel, namespace: Namespace) => {
-  switch (namespace) {
-    // Main namespaces
-    case Namespace.uniprotkb:
-      return uniProtKbConverter(row as UniProtkbAPIModel);
-    case Namespace.uniref:
-      return row as UniRefLiteAPIModel;
-    case Namespace.uniparc:
-      return row as UniParcAPIModel;
-    case Namespace.proteomes:
-      return row as ProteomesAPIModel;
-    // Supporting data
-    case Namespace.taxonomy:
-      return row as TaxonomyAPIModel;
-    case Namespace.keywords:
-      return row as KeywordsAPIModel;
-    case Namespace.citations:
-      return row as CitationsAPIModel;
-    case Namespace.diseases:
-      return row as DiseasesAPIModel;
-    case Namespace.database:
-      return row as DatabaseAPIModel;
-    case Namespace.locations:
-      return row as LocationsAPIModel;
-    default:
-      // eslint-disable-next-line no-console
-      console.warn(`Unrecognised namespace: "${namespace}"`);
-      return null;
-  }
-};
-
-// TODO: create a "Column" type to cover the different column types
-// and a Column renderer type with label: string and a render definition.
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const ColumnConfigurations: Partial<Record<Namespace, Map<any, any>>> = {
-  [Namespace.uniprotkb]: UniProtKBColumnConfiguration,
-  [Namespace.uniref]: UniRefColumnConfiguration,
-  [Namespace.uniparc]: UniParcColumnConfiguration,
-  [Namespace.proteomes]: ProteomesColumnConfiguration,
-  [Namespace.taxonomy]: TaxonomyColumnConfiguration,
-  [Namespace.keywords]: KeywordsColumnConfiguration,
-  [Namespace.citations]: CitationsColumnConfiguration,
-  [Namespace.diseases]: DiseasesColumnConfiguration,
-  [Namespace.database]: DatabaseColumnConfiguration,
-  [Namespace.locations]: LocationsColumnConfiguration,
-};
 
 const cardRenderer = (
   namespace: Namespace,
@@ -217,47 +154,6 @@ const cardRenderer = (
   }
 };
 
-type ColumnDescriptor = {
-  name: string;
-  label: ReactNode;
-  render: (row: APIModel) => ReactNode;
-  sortable?: true;
-  sorted?: SortDirection;
-};
-const getColumnsToDisplay = (
-  namespace: Namespace,
-  columns: Column[] | undefined,
-  sortableColumnToSortColumn: Map<Column, string>,
-  sortColumn: SortableColumn,
-  sortDirection: SortDirection
-): ColumnDescriptor[] =>
-  columns?.map((columnName) => {
-    const columnConfig = ColumnConfigurations[namespace]?.get(columnName);
-    if (columnConfig) {
-      const columnDescriptor = {
-        label: columnConfig.label,
-        name: columnName,
-        render: (row: APIModel) =>
-          columnConfig.render(convertRow(row, namespace)),
-      };
-      if (sortableColumnToSortColumn.has(columnName)) {
-        return {
-          ...columnDescriptor,
-          sortable: true,
-          sorted: columnName === sortColumn ? sortDirection : undefined,
-        };
-      }
-      return columnDescriptor;
-    }
-    return {
-      label: columnName,
-      name: columnName,
-      render: () => (
-        <div className="warning">{`${columnName} has no config yet`}</div>
-      ),
-    };
-  }) || [];
-
 type ResultsTableProps = {
   selectedEntries: string[];
   handleEntrySelection: (rowId: string) => void;
@@ -269,25 +165,24 @@ const ResultsView: FC<ResultsTableProps> = ({
 }) => {
   const namespace = useNS() || Namespace.uniprotkb;
   const [viewMode] = useUserPreferences<ViewMode>('view-mode', ViewMode.CARD);
-  const [columns] = useUserPreferences<Column[]>(
-    `table columns for ${namespace}` as const,
-    nsToDefaultColumns[namespace]
-  );
+  const history = useHistory();
+  const [columns, updateColumnSort] = useColumns();
 
   const prevNamespace = useRef<Namespace>(namespace);
   useEffect(() => {
     // will set it *after* the current render
     prevNamespace.current = namespace;
   });
-  const prevColumns = useRef<Column[] | undefined>(columns);
+  const prevColumns = useRef<ColumnDescriptor[] | undefined>(columns);
   useEffect(() => {
     // will set it *after* the current render
     prevColumns.current = columns;
   });
 
-  const history = useHistory();
-
-  let queryColumns = viewMode === ViewMode.CARD ? undefined : columns;
+  let queryColumns =
+    viewMode === ViewMode.CARD
+      ? undefined
+      : columns.map((columnDescriptor) => columnDescriptor.name as Column);
   if (viewMode === ViewMode.CARD) {
     // TODO: Do similar things for the rest of namespaces
     if (namespace === Namespace.uniprotkb) {
@@ -295,11 +190,9 @@ const ResultsView: FC<ResultsTableProps> = ({
     }
   }
 
-  const { url: initialApiUrl, sortColumn, sortDirection, direct } = useNSQuery({
+  const { url: initialApiUrl, direct } = useNSQuery({
     queryColumns,
   });
-
-  const [sortableColumnToSortColumn, updateColumnSort] = useColumnSort();
 
   const [url, setUrl] = useState(initialApiUrl);
   const [metaData, setMetaData] = useState<{
@@ -392,13 +285,7 @@ const ResultsView: FC<ResultsTableProps> = ({
         // Column view
         <DataTableWithLoader
           getIdKey={getIdKey}
-          columns={getColumnsToDisplay(
-            namespace,
-            columns,
-            sortableColumnToSortColumn,
-            sortColumn,
-            sortDirection
-          )}
+          columns={columns}
           data={allResults}
           selected={selectedEntries}
           onSelectRow={handleEntrySelection}
